+++ src/AStarContainer.h
@@ -74,7 +74,7 @@ class AStarContainer
         *  The data in this array is initialised as -1, which indicates hat there is no corresponding node for that position
         *  This must be maintained when nodes are added, removed and re-ordered in the node array
         */
-        int* map_pos;
+        short* map_pos;
 };
 
 /* This class is used to store the closed list of a* nodes
@@ -95,7 +95,7 @@ public:
 private:
     unsigned int size;
     AStarNode** nodes;
-    int* map_pos;
+    short* map_pos;
 
 };
